# services/interbank-service/docs/swagger.yaml
definitions:
...
  dto.CurrencyCode:
    enum:
    - RSD
    - EUR
    - USD
    - CHF
    - JPY
    - AUD
    - CAD
    - GBP
    type: string
    x-enum-varnames:
    - CurrencyRSD
    - CurrencyEUR
    - CurrencyUSD
    - CurrencyCHF
    - CurrencyJPY
    - CurrencyAUD
    - CurrencyCAD
    - CurrencyGBP
  dto.ForeignBankId:
    properties:
      id:
        maxLength: 64
        type: string
      routingNumber:
        type: integer
    required:
    - id
    - routingNumber
    type: object
...
  dto.MonetaryValue:
    properties:
      amount:
        type: number
      currency:
        $ref: '#/definitions/dto.CurrencyCode'
    required:
    - amount
    - currency
    type: object
...
  dto.OtcNegotiation:
    properties:
      amount:
        minimum: 1
        type: integer
      buyerAccountNumber:
        description: BuyerAccountNumber is a Banka-4 extension to the protocol's OtcOffer.
        type: string
      buyerId:
        $ref: '#/definitions/dto.ForeignBankId'
      isOngoing:
        type: boolean
      lastModifiedBy:
        $ref: '#/definitions/dto.ForeignBankId'
      premium:
        $ref: '#/definitions/dto.MonetaryValue'
      pricePerUnit:
        $ref: '#/definitions/dto.MonetaryValue'
      sellerId:
        $ref: '#/definitions/dto.ForeignBankId'
      settlementDate:
        type: string
      stock:
        $ref: '#/definitions/dto.StockDescription'
    required:
    - amount
    - buyerId
    - lastModifiedBy
    - pricePerUnit
    - sellerId
    - settlementDate
    - stock
    type: object
...
  dto.OtcOffer:
    properties:
      amount:
        minimum: 1
        type: integer
      buyerAccountNumber:
        description: BuyerAccountNumber is a Banka-4 extension to the protocol's OtcOffer.
        type: string
      buyerId:
        $ref: '#/definitions/dto.ForeignBankId'
      lastModifiedBy:
        $ref: '#/definitions/dto.ForeignBankId'
      premium:
        $ref: '#/definitions/dto.MonetaryValue'
      pricePerUnit:
        $ref: '#/definitions/dto.MonetaryValue'
      sellerId:
        $ref: '#/definitions/dto.ForeignBankId'
      settlementDate:
        type: string
      stock:
        $ref: '#/definitions/dto.StockDescription'
    required:
    - amount
    - buyerId
    - lastModifiedBy
    - pricePerUnit
    - sellerId
    - settlementDate
    - stock
    type: object
  dto.PeerContract:
    properties:
      amount:
        type: integer
      buyerId:
        $ref: '#/definitions/dto.ForeignBankId'
      createdAt:
        type: string
      exercisedAt:
        type: string
      id:
        $ref: '#/definitions/dto.ForeignBankId'
      negotiationId:
        $ref: '#/definitions/dto.ForeignBankId'
      premium:
        $ref: '#/definitions/dto.MonetaryValue'
      sellerId:
        $ref: '#/definitions/dto.ForeignBankId'
      settlementDate:
        type: string
      status:
        type: string
      strikePrice:
        $ref: '#/definitions/dto.MonetaryValue'
      ticker:
        type: string
      updatedAt:
        type: string
    type: object
...
  dto.PublicStock:
    properties:
      sellers:
        items:
          $ref: '#/definitions/dto.PublicStockSeller'
        type: array
      stock:
        $ref: '#/definitions/dto.StockDescription'
    type: object
  dto.PublicStockSeller:
    properties:
      amount:
        type: integer
      seller:
        $ref: '#/definitions/dto.ForeignBankId'
    type: object
  dto.StockDescription:
    properties:
      ticker:
        type: string
    required:
    - ticker
    type: object
...
  dto.UserInformation:
    properties:
      bankDisplayName:
        type: string
      displayName:
        type: string
    type: object
...
  errors.AppError:
    properties:
      message:
        type: string
      status:
        type: string
      timestamp:
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    Bank-to-bank coordination service implementing the inter-bank
    transaction and OTC negotiation protocol. Authentication is
    per-peer via the X-Api-Key header (see §2.10).
  title: Interbank Service API
  version: "1.0"
paths:
...
  /interbank/negotiations:
    post:
      consumes:
      - application/json
      description: |-
        §3.2 — peer bank's buyer initiates a negotiation against
        a seller in our bank. Returns the ForeignBankId of the
        new negotiation owned by us.
      parameters:
      - description: Peer bank API key
        in: header
        name: X-Api-Key
        required: true
        type: string
      - description: Initial offer
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.OtcOffer'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ForeignBankId'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
      summary: Create OTC negotiation (peer-initiated)
      tags:
      - interbank-otc
  /interbank/negotiations/{rn}/{id}:
    delete:
      description: |-
        §3.5 — either party may withdraw from a negotiation (the
        recipient may be the authoritative or the mirror side). The
        operation is idempotent: closing an already-closed negotiation
        returns 204 without changing state.
      parameters:
      - description: Peer bank API key
        in: header
        name: X-Api-Key
        required: true
        type: string
      - description: Authoritative (seller's) routing number
        in: path
        name: rn
        required: true
        type: integer
      - description: Negotiation id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.AppError'
      summary: Close OTC negotiation
      tags:
      - interbank-otc
    get:
      description: |-
        §3.4 — returns the current negotiation state as the spec shape
        (the OtcOffer fields plus isOngoing). The :rn path parameter is
        the negotiation's authoritative (seller's) routing number.
      parameters:
      - description: Peer bank API key
        in: header
        name: X-Api-Key
        required: true
        type: string
      - description: Authoritative (seller's) routing number
        in: path
        name: rn
        required: true
        type: integer
      - description: Negotiation id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.OtcNegotiation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.AppError'
      summary: Get OTC negotiation
      tags:
      - interbank-otc
    put:
      consumes:
      - application/json
      description: |-
        §3.3 — a peer bank posts a counter-offer (either party may
        counter; the recipient may be the authoritative seller's bank
        or the buyer's mirror-holding bank). Buyer/seller and ticker are
        immutable; only negotiable parameters may change. The same party
        may not counter twice in a row (turn rule).
      parameters:
      - description: Peer bank API key
        in: header
        name: X-Api-Key
        required: true
        type: string
      - description: Authoritative (seller's) routing number
        in: path
        name: rn
        required: true
        type: integer
      - description: Negotiation id
        in: path
        name: id
        required: true
        type: string
      - description: Updated offer
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.OtcOffer'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.AppError'
        "409":
          description: Turn violation or negotiation closed
          schema:
            $ref: '#/definitions/errors.AppError'
      summary: Post counter-offer
      tags:
      - interbank-otc
  /interbank/negotiations/{rn}/{id}/accept:
    get:
      description: |-
        §3.6 — accepts the current offer on a negotiation owned by us
        (we are the seller's authoritative bank). Forms the option
        contract and drives the §2 NEW_TX that moves the premium, then
        returns the resulting contract.
      parameters:
      - description: Peer bank API key
        in: header
        name: X-Api-Key
        required: true
        type: string
      - description: Routing number (this bank)
        in: path
        name: rn
        required: true
        type: integer
      - description: Negotiation id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PeerContract'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.AppError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/errors.AppError'
      summary: Accept OTC negotiation
      tags:
      - interbank-otc
  /interbank/public-stock:
    get:
      description: |-
        §3.1 — returns every stock holding at this bank that has
        a non-zero public amount, grouped by ticker with the list
        of sellers and their available quantities. Data pulled
        from trading-service via gRPC.
      parameters:
      - description: Peer bank API key
        in: header
        name: X-Api-Key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PublicStock'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errors.AppError'
      summary: List public stocks at this bank
      tags:
      - interbank-otc
  /interbank/user/{rn}/{id}:
    get:
      description: |-
        §3.7 — peer banks call us to resolve one of our users into
        a display name. routingNumber must match ours; id is the
        local user id encoded as a decimal string.
      parameters:
      - description: Peer bank API key
        in: header
        name: X-Api-Key
        required: true
        type: string
      - description: Routing number (this bank)
        in: path
        name: rn
        required: true
        type: integer
      - description: User id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserInformation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errors.AppError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.AppError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errors.AppError'
      summary: Resolve user display name
      tags:
      - interbank-otc
swagger: "2.0"